refactor : settings 폴더명을 constants로 변경
settings는 config 파일과 혼동될 수 있어, 상수를 의미하는 constants로 변경
