+++ __tests__/RandomNumberGenerator.test.js
@@ -1,5 +1,5 @@
 import randomNumberGenerator from "../src/model/RandomNumberGenerator.js";
-import { systemSetting } from "../src/settings/systemSetting.js";
+import { systemSetting } from "../src/constants/systemSetting.js";
 test("randomNumberGenerator는 0~9 사이 수를 생성해야 한다", () => {
   //given
   const number = randomNumberGenerator(
