+++ src/service/raceService.js
@@ -3,7 +3,7 @@ import { InputView } from "../view/InputView.js";
 import { OutputView } from "../view/OutputView.js";
 import { parsingService } from "./parsingService.js";
 import randomNumberGenerator from "../model/RandomNumberGenerator.js";
-import { systemSetting } from "../settings/systemSetting.js";
+import { systemSetting } from "../constants/systemSetting.js";
 
 export async function raceInit() {
   const cars = [];
